Actions2Methods: GetPermissions, HasPermission, SetPermisssions.
