+++ src/Tests/SenseNet.ODataTests/ODataOperationTests.cs
@@ -12,7 +12,6 @@ using SenseNet.ContentRepository;
 using SenseNet.ContentRepository.Schema;
 using SenseNet.ContentRepository.Storage;
 using SenseNet.OData;
-using SenseNet.OData.Operations;
 using SenseNet.Portal.ApplicationModel;
 using SenseNet.Search;
 using Task = System.Threading.Tasks.Task;
@@ -754,9 +753,9 @@ namespace SenseNet.ODataTests
                     case "Action2": return new Action2();
                     case "Action3": return new Action3();
                     case "Action4": return new Action4();
-                    case "GetPermissions": return new GetPermissionsAction();
-                    case "SetPermissions": return new SetPermissionsAction();
-                    case "HasPermission": return new HasPermissionAction();
+                    //case "GetPermissions": return new GetPermissionsAction();
+                    //case "SetPermissions": return new SetPermissionsAction();
+                    //case "HasPermission": return new HasPermissionAction();
                     //case "AddAspects": return new SenseNet.ApplicationModel.AspectActions.AddAspectsAction();
                     //case "RemoveAspects": return new SenseNet.ApplicationModel.AspectActions.RemoveAspectsAction();
                     //case "RemoveAllAspects": return new SenseNet.ApplicationModel.AspectActions.RemoveAllAspectsAction();
